Dump of previous work and code examples
This contains a wide selection of code I've written. Some of it I wrote
as part of
projects at university; some of it I wrote for fun; and some of it I
wrote so I didn't forget how to code after finishing uni before getting
a job. ;-)

Note that TollBridge, Fibonacci and atoi_ex are QtCreator projects and
therefore contain the plethora of fairly uninteresting files that comes
with that.

As this repo is designed to be used just as a place to dump my old code
to make it easier to store, access and distribute to anyone who might
want to see it, I don't particularly care for removing the non-necessary
files.