Labyrinth: Final year uni programming project
This is an OpenGL based game that I developed as part of a team or 3;
the other team members being Marc DiLuzio and Gary Nisbet; at
Bournemouth University on the Computer Visualisation and Animation
course of 2012.

It is important that this code and its accompanying files, including but
not limited to any sound, texture, or object files, is never used wholly
or in part for commercial use as technically it "belongs" to the
university. Any work that wasn't produced by us is attributed to the
contents owner in the documentation.

Please note that we are no longer working on this project - any use of
this project is entirely unsupported and you are on your own. If you
have queries regarding the code content, or any of the supporting files
I will try to answer your questions, however neither I nor Marc or Gary
will be further developing or maintaining this project.